Header tweaking for compilation speed-up

+++ src/osgEarthDrivers/engine_osgterrain/SinglePassTerrainTechnique.cpp
@@ -943,19 +943,19 @@ SinglePassTerrainTechnique::createGeometry( const TileFrame& tilef )
     {
       int min_i = (int)floor((*mr)._ndcMin.x() * (double)(numColumns-1));
       if (min_i < 0) min_i = 0;
-      if (min_i >= numColumns) min_i = numColumns - 1;
+      if (min_i >= (int)numColumns) min_i = numColumns - 1;
 
       int min_j = (int)floor((*mr)._ndcMin.y() * (double)(numRows-1));
       if (min_j < 0) min_j = 0;
-      if (min_j >= numColumns) min_j = numColumns - 1;
+      if (min_j >= (int)numColumns) min_j = numColumns - 1;
 
       int max_i = (int)ceil((*mr)._ndcMax.x() * (double)(numColumns-1));
       if (max_i < 0) max_i = 0;
-      if (max_i >= numColumns) max_i = numColumns - 1;
+      if (max_i >= (int)numColumns) max_i = numColumns - 1;
 
       int max_j = (int)ceil((*mr)._ndcMax.y() * (double)(numRows-1));
       if (max_j < 0) max_j = 0;
-      if (max_j >= numColumns) max_j = numColumns - 1;
+      if (max_j >= (int)numColumns) max_j = numColumns - 1;
 
       if (min_i >= 0 && max_i >= 0 && min_j >= 0 && max_j >= 0)
       {
@@ -1779,7 +1779,7 @@ SinglePassTerrainTechnique::createGeometry( const TileFrame& tilef )
 
         //Add a primative set for each continuous skirt strip
         skirtBreaks.push_back(skirtVerts->size());
-        for (int p=1; p < skirtBreaks.size(); p++)
+        for (unsigned p=1; p < skirtBreaks.size(); p++)
           skirt->addPrimitiveSet( new osg::DrawArrays( GL_TRIANGLE_STRIP, skirtBreaks[p-1], skirtBreaks[p] - skirtBreaks[p-1] ) );
     }
